Rename JreCaching to JreResolution, fix namespaces
