+++ Tests/SonarScanner.MSBuild.PreProcessor.Test/JreCaching/JreCacheTests.cs
@@ -28,7 +28,10 @@ using NSubstitute;
 using NSubstitute.ExceptionExtensions;
 using NSubstitute.ReturnsExtensions;
 using SonarScanner.MSBuild.Common;
+using SonarScanner.MSBuild.PreProcessor.Interfaces;
 using SonarScanner.MSBuild.PreProcessor.JreCaching;
+using SonarScanner.MSBuild.PreProcessor.JreResolution;
+using SonarScanner.MSBuild.PreProcessor.Unpacking;
 using TestUtilities;
 
 namespace SonarScanner.MSBuild.PreProcessor.Test.JreCaching;
