+++ Tests/SonarScanner.MSBuild.PreProcessor.Test/JreCaching/JreResolverTests.cs
@@ -26,6 +26,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NSubstitute;
 using SonarScanner.MSBuild.Common;
 using SonarScanner.MSBuild.PreProcessor.JreCaching;
+using SonarScanner.MSBuild.PreProcessor.JreResolution;
 using TestUtilities;
 
 namespace SonarScanner.MSBuild.PreProcessor.Test.JreCaching;
