+++ Tests/SonarScanner.MSBuild.PreProcessor.Test/JreCaching/ZipUnpackTests.cs
@@ -22,7 +22,7 @@ using System;
 using System.IO;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SonarScanner.MSBuild.PreProcessor.JreCaching;
+using SonarScanner.MSBuild.PreProcessor.Unpacking;
 using TestUtilities;
 
 namespace SonarScanner.MSBuild.PreProcessor.Test.JreCaching;
