+++ src/SonarScanner.MSBuild.PreProcessor/ChecksumSha256.cs
@@ -18,11 +18,21 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using SonarScanner.MSBuild.Common;
+using System;
+using System.IO;
+using System.Security.Cryptography;
+using SonarScanner.MSBuild.PreProcessor.Interfaces;
 
 namespace SonarScanner.MSBuild.PreProcessor.JreCaching;
 
-public interface IUnpackerFactory
+internal class ChecksumSha256 : IChecksum
 {
-    IUnpacker Create(ILogger logger, IDirectoryWrapper directoryWrapper, IFileWrapper fileWrapper, IFilePermissionsWrapper filePermissionsWrapper, string archivePath);
+    public static ChecksumSha256 Instance { get; } = new();
+
+    public string ComputeHash(Stream source)
+    {
+        using var sha = SHA256.Create();
+        var bytes = sha.ComputeHash(source);
+        return BitConverter.ToString(bytes).Replace("-", string.Empty).ToLowerInvariant();
+    }
 }
