+++ src/SonarScanner.MSBuild.PreProcessor/Interfaces/IChecksum.cs
@@ -18,18 +18,11 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using SonarScanner.MSBuild.PreProcessor.JreCaching;
+using System.IO;
 
-namespace SonarScanner.MSBuild.PreProcessor;
+namespace SonarScanner.MSBuild.PreProcessor.Interfaces;
 
-public sealed class JreMetadata(string Id, string Filename, string JavaPath, string DownloadUrl, string Sha256)
+public interface IChecksum
 {
-    public string Id { get; } = Id;                     // Optional, only exists for SonarQube
-    public string Filename { get; } = Filename;
-    public string Sha256 { get; } = Sha256;
-    public string JavaPath { get; } = JavaPath;
-    public string DownloadUrl { get; } = DownloadUrl;   // Optional, only exists for SonarCloud
-
-    public JreDescriptor ToDescriptor() =>
-        new(Filename, Sha256, JavaPath);
+    string ComputeHash(Stream source);
 }
