+++ src/SonarScanner.MSBuild.PreProcessor/JreResolution/IJreCache.cs
@@ -18,11 +18,15 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using ICSharpCode.SharpZipLib.Tar;
+using System;
+using System.IO;
+using System.Threading.Tasks;
 
-namespace SonarScanner.MSBuild.PreProcessor.JreCaching;
+namespace SonarScanner.MSBuild.PreProcessor.JreResolution;
 
-public interface IFilePermissionsWrapper
+public interface IJreCache
 {
-    void Copy(TarEntry sourceEntry, string destinationPath);
+    JreCacheResult IsJreCached(string sonarUserHome, JreDescriptor jreDescriptor);
+
+    Task<JreCacheResult> DownloadJreAsync(string sonarUserHome, JreDescriptor jreDescriptor, Func<Task<Stream>> jreDownload);
 }
